+++ quadtrees/Vector2d.h
@@ -9,4 +9,4 @@ double length(const Vector2d& vec);
 Vector2d distance(const Vector2d& vec1, const Vector2d& vec2);
 Vector2d unit(const Vector2d& vector);
 Vector2d sum(const Vector2d& vec1, const Vector2d& vec2);
-Vector2d divide(const Vector2d& vec, const int value);
+Vector2d divide(const Vector2d& vec, const double value);
